refactor
update imports paths
remove unused `getEnumValue`
add ShareProvider to sharebutton-class.ts
