+++ src/helpers/index.ts
@@ -2,4 +2,3 @@ export * from './n-formatter.pipe';
 export * from './share.helper';
 export * from './share-buttons.class';
 export * from './share-links.functions';
-export * from './share-provider.enum';
