+++ src/helpers/share-buttons.class.ts
@@ -1,8 +1,18 @@
-import { ShareProvider } from './share-provider.enum';
+
+export const ShareProvider = {
+    FACEBOOK: 'facebook',
+    TWITTER: 'twitter',
+    REDDIT: 'reddit',
+    STUMBLEUPON: 'stumble',
+    LINKEDIN: 'linkedin',
+    GOOGLEPLUS: 'google',
+    TUMBLR: 'tumblr',
+    PINTEREST: 'pinterest'
+};
 
 export class ShareButton {
 
-    constructor(public provider: ShareProvider,
+    constructor(public provider: string,
         public template: string,
         public classes: string) {
     }
