+++ src/helpers/share.helper.spec.ts
@@ -90,36 +90,4 @@ describe('Module: Helper, Isolate Tests', () => {
 
     });
 
-    describe('getEnumValue()', () => {
-
-        it('should return undefined from invalid enum string literal', () => {
-            expect(Helper.getEnumValue('invalid', ShareProvider)).toBeUndefined();
-        });
-
-        it('should return undefined from invalid enum numeric value', () => {
-            expect(Helper.getEnumValue(-1, ShareProvider)).toBeUndefined();
-            expect(Helper.getEnumValue(999, ShareProvider)).toBeUndefined();
-        });
-
-        it('should return enum value from valid string literal (case insensitive)', () => {
-            expect(Helper.getEnumValue('linkedin', ShareProvider)).toEqual(ShareProvider.LINKEDIN);
-            expect(Helper.getEnumValue('LINKEDIN', ShareProvider)).toEqual(ShareProvider.LINKEDIN);
-            expect(Helper.getEnumValue('LinkedIn', ShareProvider)).toEqual(ShareProvider.LINKEDIN);
-        });
-
-        it('should return enum value from valid numeric value', () => {
-            expect(Helper.getEnumValue(0, ShareProvider)).toEqual(ShareProvider.FACEBOOK);
-            expect(Helper.getEnumValue(3, ShareProvider)).toEqual(ShareProvider.STUMBLEUPON);
-            expect(Helper.getEnumValue(4, ShareProvider)).toEqual(ShareProvider.LINKEDIN);
-            expect(Helper.getEnumValue(5, ShareProvider)).toEqual(ShareProvider.GOOGLEPLUS);
-        });
-
-        it('should return enum value from valid numeric value', () => {
-
-            expect(Helper.getEnumValue(0, ShareProvider)).toEqual(ShareProvider.FACEBOOK);
-        });
-
-    });
-
-
 });
